Add tests for functions generating Cypher statements
