+++ concepts/concepts_service.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/Financial-Times/concepts-rw-neo4j/ontology"
@@ -181,207 +180,6 @@ func (s *ConceptService) read(uuid string, transID string) (ontology.NewAggregat
 	}
 
 	return newAggregatedConcept, true, nil
-}
-
-func getReadStatement() string {
-	statementTemplate := `
-		MATCH (canonical:Thing {prefUUID:{uuid}})<-[:EQUIVALENT_TO]-(source:Thing)
-		OPTIONAL MATCH (source)-[:ISSUED_BY]->(issuer:Thing)
-		%s
-		WITH
-			canonical,
-			issuer,
-			source,
-			%s
-			ORDER BY
-				source.uuid,
-				hasRoleNode.uuid
-		WITH
-			canonical,
-			issuer,
-			hasMemberNode,
-			hasOrganisationNode,
-			{
-				uuid: source.uuid,
-				prefLabel: source.prefLabel,
-				types: labels(source),
-				authority: source.authority,
-				authorityValue: source.authorityValue,
-				figiCode: source.figiCode,
-				lastModifiedEpoch: source.lastModifiedEpoch,
-				isDeprecated: source.isDeprecated,
-				industryIdentifier: source.industryIdentifier,
-				issuedBy: issuer.uuid,
-				%s
-			} as sources,
-			collect({
-				inceptionDate: hasRoleRel.inceptionDate,
-				inceptionDateEpoch: hasRoleRel.inceptionDateEpoch,
-				membershipRoleUUID: hasRoleNode.uuid,
-				terminationDate: hasRoleRel.terminationDate,
-				terminationDateEpoch: hasRoleRel.terminationDateEpoch
-			}) as membershipRoles
-		RETURN
-			canonical.aggregateHash as aggregateHash,
-			canonical.aliases as aliases,
-			canonical.descriptionXML as descriptionXML,
-			canonical.emailAddress as emailAddress,
-			canonical.facebookPage as facebookPage,
-			canonical.figiCode as figiCode,
-			canonical.imageUrl as imageUrl,
-			canonical.inceptionDate as inceptionDate,
-			canonical.inceptionDateEpoch as inceptionDateEpoch,
-			canonical.prefLabel as prefLabel,
-			canonical.prefUUID as prefUUID,
-			canonical.scopeNote as scopeNote,
-			canonical.shortLabel as shortLabel,
-			canonical.strapline as strapline,
-			canonical.terminationDate as terminationDate,
-			canonical.terminationDateEpoch as terminationDateEpoch,
-			canonical.twitterHandle as twitterHandle,
-			canonical.properName as properName,
-			canonical.shortName as shortName,
-			canonical.tradeNames as tradeNames,
-			canonical.formerNames as formerNames,
-			canonical.countryCode as countryCode,
-			canonical.countryOfIncorporation as countryOfIncorporation,
-			canonical.countryOfOperations as countryOfOperations,
-			canonical.countryOfRisk as countryOfRisk,
-			canonical.postalCode as postalCode,
-			canonical.yearFounded as yearFounded,
-			canonical.leiCode as leiCode,
-			canonical.isDeprecated as isDeprecated,
-			canonical.salutation as salutation,
-			canonical.birthYear as birthYear,
-			canonical.iso31661 as iso31661,
-			canonical.industryIdentifier as industryIdentifier,
-			issuer.uuid as issuedBy,
-			hasOrganisationNode.uuid as organisationUUID,
-			hasMemberNode.uuid as personUUID,
-			membershipRoles,
-			labels(canonical) as types,
-			collect(sources) as sourceRepresentations`
-
-	return fmt.Sprintf(statementTemplate,
-		strings.Join(getOptionalMatches(), "\n"),
-		strings.Join(getWithMatched(), ",\n"),
-		strings.Join(getSourceRels(), ",\n"))
-}
-
-func getOptionalMatches() []string {
-	var relOptionalMatches []string
-	for relLabel, relCfg := range ontology.GetConfig().Relationships {
-		relOptionalMatches = append(relOptionalMatches, getOptionalMatch(relLabel, relCfg))
-	}
-
-	for relLabel, relCfg := range relationships {
-		relOptionalMatches = append(relOptionalMatches, getOptionalMatch(relLabel, relCfg))
-	}
-
-	return relOptionalMatches
-}
-
-func getWithMatched() []string {
-	var withMatched []string
-	for relLabel, relCfg := range ontology.GetConfig().Relationships {
-		withMatched = append(withMatched, getMatched(relLabel, relCfg)...)
-	}
-
-	for relLabel, relCfg := range relationships {
-		withMatched = append(withMatched, getMatched(relLabel, relCfg)...)
-	}
-
-	return withMatched
-}
-
-func getSourceRels() []string {
-	var sourceRels []string
-	for relLabel, relCfg := range ontology.GetConfig().Relationships {
-		sourceRels = append(sourceRels, getSourceRel(relLabel, relCfg))
-	}
-
-	for relLabel, relCfg := range relationships {
-		sourceRels = append(sourceRels, getSourceRel(relLabel, relCfg))
-	}
-
-	return sourceRels
-}
-
-func getOptionalMatch(relLabel string, relCfg ontology.RelationshipConfig) string {
-	r := toCamelCase(relLabel)
-	nodeName := r + "Node"
-
-	var relName string
-	if len(relCfg.Properties) > 0 {
-		relName = r + "Rel"
-	}
-
-	toNodeLabel := "Thing"
-	if relCfg.ToNodeWithLabel != "" {
-		toNodeLabel = relCfg.ToNodeWithLabel
-	}
-
-	return fmt.Sprintf("OPTIONAL MATCH (source)-[%s:%s]->(%s:%s)", relName, relLabel, nodeName, toNodeLabel)
-}
-
-func getMatched(relLabel string, relCfg ontology.RelationshipConfig) []string {
-	r := toCamelCase(relLabel)
-	nodeName := r + "Node"
-
-	var relName string
-	if len(relCfg.Properties) > 0 {
-		relName = r + "Rel"
-	}
-
-	matched := []string{nodeName}
-	if len(relCfg.Properties) > 0 {
-		matched = append(matched, relName)
-	}
-
-	return matched
-}
-
-func getSourceRel(relLabel string, relCfg ontology.RelationshipConfig) string {
-	r := toCamelCase(relLabel)
-	nodeName := r + "Node"
-
-	var relName string
-	if len(relCfg.Properties) > 0 {
-		relName = r + "Rel"
-	}
-
-	if relCfg.OneToOne {
-		return fmt.Sprintf("%s: %s.uuid", relCfg.ConceptField, nodeName)
-	} else if len(relCfg.Properties) == 0 {
-		return fmt.Sprintf("%s: collect(DISTINCT %s.uuid)", relCfg.ConceptField, nodeName)
-	} else {
-		var relProps []string
-		for _, relProp := range relCfg.Properties {
-			relProps = append(relProps, fmt.Sprintf("%s: %s.%s", relProp, relName, relProp))
-		}
-
-		uuidField := "UUID"
-		if relLabel == "HAS_ROLE" {
-			uuidField = "membershipRoleUUID"
-		}
-
-		return fmt.Sprintf("%s: collect(DISTINCT {%s: %s.uuid, %s})",
-			relCfg.ConceptField,
-			uuidField,
-			nodeName,
-			strings.Join(relProps, ", "))
-	}
-}
-
-func toCamelCase(relLabel string) string {
-	labelLower := strings.ToLower(relLabel)
-	var vals []string
-	for _, val := range strings.Split(labelLower, "_") {
-		vals = append(vals, strings.Title(val))
-	}
-
-	vals[0] = strings.ToLower(vals[0])
-	return strings.Join(vals, "")
 }
 
 func (s *ConceptService) Write(thing interface{}, transID string) (interface{}, error) {
@@ -816,45 +614,11 @@ func deleteLonePrefUUID(prefUUID string) *neoism.CypherQuery {
 
 //Clear down current concept node
 func (s *ConceptService) clearDownExistingNodes(ac ontology.NewAggregatedConcept) []*neoism.CypherQuery {
-	var relOptionalMatches, relNames []string
-	for relLabel := range ontology.GetConfig().Relationships {
-		r := toCamelCase(relLabel)
-		relName := r + "Rel"
-		nodeName := r + "Node"
-		relOptionalMatch := fmt.Sprintf("OPTIONAL MATCH (t)-[%s:%s]->(%s)", relName, relLabel, nodeName)
-
-		relOptionalMatches = append(relOptionalMatches, relOptionalMatch)
-		relNames = append(relNames, relName)
-	}
-
-	for relLabel := range relationships {
-		r := toCamelCase(relLabel)
-		relName := r + "Rel"
-		nodeName := r + "Node"
-		relOptionalMatch := fmt.Sprintf("OPTIONAL MATCH (t)-[%s:%s]->(%s)", relName, relLabel, nodeName)
-
-		relOptionalMatches = append(relOptionalMatches, relOptionalMatch)
-		relNames = append(relNames, relName)
-	}
-
 	acUUID := ac.PrefUUID
 	var queryBatch []*neoism.CypherQuery
 	for _, sr := range ac.SourceRepresentations {
-		//nolint:gosec
-		deleteStatement := fmt.Sprintf(`
-			MATCH (t:Thing {uuid:{id}})
-			OPTIONAL MATCH (t)-[eq:EQUIVALENT_TO]->(a:Thing)
-			OPTIONAL MATCH (t)-[issuerRel:ISSUED_BY]->(issuer)
-			%s
-			REMOVE t:%s
-			SET t={uuid:{id}}
-			DELETE eq, issuerRel, %s`,
-			strings.Join(relOptionalMatches, "\n"),
-			getLabelsToRemove(),
-			strings.Join(relNames, ", "))
-
 		deletePreviousSourceLabelsAndPropertiesQuery := &neoism.CypherQuery{
-			Statement: deleteStatement,
+			Statement: getDeleteStatement(),
 			Parameters: map[string]interface{}{
 				"id": sr.UUID,
 			},
